+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# [Master][]
+
+- Fix support for `ActiveSupport::TimeWithZone`.
+
 # [0.2.1][] (2013-08-06)
 
 - Fix setting a default value on more than one attribute at a time.
@@ -27,6 +31,7 @@
 
 - Initial release.
 
+  [master]: https://github.com/orgsync/active_interaction/compare/v0.2.0...master
   [0.2.1]: https://github.com/orgsync/active_interaction/compare/v0.2.0...v0.2.1
   [0.2.0]: https://github.com/orgsync/active_interaction/compare/v0.1.3...v0.2.0
   [0.1.3]: https://github.com/orgsync/active_interaction/compare/v0.1.2...v0.1.3